chidi-backend/services/api-gateway/pyproject.toml
[tool.poetry]
name = "chidi-backend"
version = "0.1.0"
description = "FastAPI backend for Chidi project"
authors = ["Mo Sheriff <mailsheriff.mo@gmail.com>"]
readme = "README.md"
packages = [{include = "chidi_backend", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
supabase = "^2.5.0"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.9"
pydantic = {extras = ["email"], version = "^2.7.4"}
httpx = "^0.27.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.1.0"
mypy = "^1.10.1"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
